+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 - change type constraint of some builtin method
     - ``Array#sort``, ``Array#searchSorted`` require ``Ord_`` type
     - ``Array#indexOf``, ``Array#lastIndexOf``, ``Array#contains`` require ``Eq_`` type
+- fix emoji sequence handling of ``String#width`` method and line renderer
+    - now width of grapheme cluster (except for emoji zwj sequence) is sum of the code point width
+    - width of emoji variation sequence (VS16, emoji-style) is always 2
 
 ### Fixed